refactor: ChessGame 내에서 Result 생성하도록 변경
