+++ src/test/java/chess/PositionFixture.java
@@ -1,4 +1,4 @@
-package chess.domain;
+package chess;
 
 import chess.domain.board.Position;
