+++ src/test/java/chess/controller/command/EndCommandTest.java
@@ -14,7 +14,7 @@ import java.util.Arrays;
 import java.util.List;
 import java.util.stream.Collectors;
 
-import static chess.domain.PositionFixture.B_2;
+import static chess.PositionFixture.B_2;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
