+++ src/test/java/chess/domain/piece/EmptyTest.java
@@ -7,7 +7,7 @@ import org.junit.jupiter.api.DisplayNameGeneration;
 import org.junit.jupiter.api.DisplayNameGenerator;
 import org.junit.jupiter.api.Test;
 
-import static chess.domain.PositionFixture.C_4;
+import static chess.PositionFixture.C_4;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
